+++ src/main/java/fr/insee/protools/backend/service/platine/pilotage/PlatinePilotageService.java
@@ -1,12 +1,16 @@
 package fr.insee.protools.backend.service.platine.pilotage;
 
+import fr.insee.protools.backend.service.platine.pilotage.dto.PlatinePilotageEligibleDto;
+import fr.insee.protools.backend.service.platine.pilotage.dto.contact.PlatineContactDto;
 import fr.insee.protools.backend.service.platine.pilotage.dto.query.QuestioningWebclientDto;
 import fr.insee.protools.backend.service.platine.pilotage.metadata.MetadataDto;
 import fr.insee.protools.backend.webclient.WebClientHelper;
 import lombok.extern.slf4j.Slf4j;
+import org.slf4j.event.Level;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import static fr.insee.protools.backend.webclient.WebClientHelper.logJson;
 import static fr.insee.protools.backend.webclient.configuration.ApiConfigProperties.KNOWN_API.KNOWN_API_PLATINE_PILOTAGE;
 
 @Service
@@ -16,7 +20,7 @@ public class PlatinePilotageService {
     @Autowired WebClientHelper webClientHelper;
     public void putMetadata(String partitionId , MetadataDto dto) {
         log.debug("putMetadata : partitionId={} - dto.su.id={} ",partitionId,dto.getSurveyDto().getId());
-        WebClientHelper.logDebugJson(String.format("putMetadata - partitionId=%s : ",partitionId),dto);
+        logJson(String.format("putMetadata - partitionId=%s : ",partitionId),dto, log,Level.DEBUG);
         var response = webClientHelper.getWebClient(KNOWN_API_PLATINE_PILOTAGE)
                 .put()
                 .uri(uriBuilder -> uriBuilder
@@ -30,7 +34,7 @@ public class PlatinePilotageService {
     }
     public void putQuestionings(QuestioningWebclientDto dto) {
         log.debug("putQuestionings: idPartitioning={} - idSu={}",dto.getIdPartitioning(),dto.getSurveyUnit().getIdSu());
-        WebClientHelper.logDebugJson("putQuestionings ",dto);
+        logJson("putMetadata ",dto,log,Level.TRACE);
         var response = webClientHelper.getWebClient(KNOWN_API_PLATINE_PILOTAGE)
                 .put()
                 .uri("/api/questionings")
@@ -41,4 +45,47 @@ public class PlatinePilotageService {
         log.trace("putQuestionings - response={} ",response);
     }
 
+    public PlatineContactDto getSUMainContact(Long idSU, String platinePartitionId){
+        log.debug("getSUMainContact: platinePartitionId={} - idSu={}",platinePartitionId,idSU);
+        PlatineContactDto response = webClientHelper.getWebClient(KNOWN_API_PLATINE_PILOTAGE)
+                .get()
+                .uri(uriBuilder -> uriBuilder
+                        .path("/api/main-contact")
+                        .queryParam("partitioning", platinePartitionId)
+                        .queryParam("survey-unit", idSU)
+                        .build())
+                .retrieve()
+                .bodyToMono(PlatineContactDto.class)
+                .block();
+        logJson("getSUMainContact response : ",response,log,Level.TRACE);
+        return response;
+    }
+
+    public Boolean isToFollowUp(Long idSU, String platinePartitionId){
+        log.debug("isToFollowUp: platinePartitionId={} - idSu={}",platinePartitionId,idSU);
+        PlatinePilotageEligibleDto response = webClientHelper.getWebClient(KNOWN_API_PLATINE_PILOTAGE)
+                .get()
+                .uri(uriBuilder -> uriBuilder
+                        .path("/api/partitionings/{idPartitioning}/survey-units/{idSu}/follow-up")
+                        .build(platinePartitionId,idSU))
+                .retrieve()
+                .bodyToMono(PlatinePilotageEligibleDto.class)
+                .block();
+        Boolean result =  Boolean.valueOf(response.getEligible());
+        logJson("isToFollowUp: result="+result+" -  response : ",response,log,Level.TRACE);
+        return  result;
+    }
+
+    public void addFollowUpState(Long idSU, String platinePartitionId){
+        log.debug("isToFollowUp: platinePartitionId={} - idSu={}",platinePartitionId,idSU);
+        String response = webClientHelper.getWebClient(KNOWN_API_PLATINE_PILOTAGE)
+                .post()
+                .uri(uriBuilder -> uriBuilder
+                        .path("/api/partitionings/{idPartitioning}/survey-units/{idSu}/follow-up")
+                        .build(platinePartitionId,idSU))
+                .retrieve()
+                .bodyToMono(String.class)
+                .block();
+        log.debug("isToFollowUp: platinePartitionId={} - idSu={} - response={}",platinePartitionId,idSU,response);
+    }
 }